+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Next
 
+- Add `cellSize` to define the size of the cell (#136)
 - Add `splitAll()` to scatter all piles (#144)
 - Add `pileBy()` for layout-, location-, and data-driven piling (#129)
 - Add `pileItemOrder` to sort the items on a pile by a callback function